переделал иницализация пользывателя и карточек для правильного отображения кнопки удалить

+++ src/components/cards.js
@@ -2,20 +2,22 @@ const templateCard = document.querySelector("#card-template").content.querySelec
 
 /* ФУНКЦИИ */
 /*Функция создания карточки*/
-function createCard(cardData, deleteCard, checkLike, openImgPopup, myUserId) {
+function createCard(cardData, deleteCard, checkLike, openImgPopup, userData) {
   const card = templateCard.cloneNode(true);
   const deleteButton = card.querySelector(".card__delete-button");
   const likeButton = card.querySelector(".card__like-button");
   const image = card.querySelector(".card__image");
   const numberLike = card.querySelector('.card__number-like')
-  checkDelete(deleteButton, myUserId, cardData.owner._id);
   card.querySelector(".card__title").textContent = cardData.name;
   image.setAttribute("src", cardData.link);
   image.setAttribute("alt", cardData.name);
   numberLike.textContent = Object.keys(cardData.likes).length
-  deleteButton.addEventListener("click", () => {
-    deleteCard(card);
-  });
+  if (userData._id === cardData.owner._id) {
+    deleteButton.classList.add('card__delete-button-visible')
+    deleteButton.addEventListener("click", () => {
+      deleteCard(card);
+    });
+  };
   likeButton.addEventListener("click", checkLike);
   image.addEventListener("click", () => {
       openImgPopup(cardData.name, cardData.link)
@@ -33,12 +35,6 @@ function deleteCard(cardElement) {
 function checkLike(evt) {
   evt.target.classList.toggle("card__like-button_is-active");
 }
-
-function checkDelete (deleteButton, myUserId, userId) {
-  if (!myUserId === userId) {
-    deleteButton.remove()
-  }
-};
 
 /* ЭКСПОРТ МОДУЛЕЙ */
 export {
